requirements.txt
Package           Version
----------------- -----------
attrs             23.2.0
boto3             1.34.111
botocore          1.34.111
certifi           2024.2.2
exceptiongroup    1.2.1
h11               0.14.0
idna              3.7
jmespath          1.0.1
numpy             1.26.4
outcome           1.3.0.post0
pip               22.0.2
PySocks           1.7.1
python-dateutil   2.9.0.post0
python-dotenv     1.0.1
pytz              2024.1
s3transfer        0.10.1
selenium          4.21.0
setuptools        59.6.0
six               1.16.0
sniffio           1.3.1
sortedcontainers  2.4.0
trio              0.25.1
trio-websocket    0.11.1
typing_extensions 4.11.0
tzdata            2024.1
urllib3           2.2.1
wsproto           1.2.0
